Eliminate the getregref_ ops.
These meant that we could not rely on ->work no longer being needed
after the frame's execution was completed. This implied that, short of
static analysis to check these ops were not used, we could not be sure
it was OK to throw out ->work right after the frame's dynamic scope
ended. The ops were not being used in Rakudo, nor in NQP besides in
the code that compiled into them, which was already been removed. They
only existed to allow lowering of lexicalref => localref at some point
in the future, but an optimization that pessimizes the world is hardly
a win. It would also have been a huge pain for the JVM backend, and
likely the JS backend too, to implement localref (neither did so).