Re-order/simplify operations in remove_one_frame.
If the frame is on the call stack, we don't need to NULL stuff in it;
it's going away anyway. We also eliminate a duplicate call to the
function MVM_args_proc_cleanup_for_cache for the callstack case (the
MVM_args_proc_cleanup in MVM_frame_destory calls it anyway). We may
have continuation tags in a frame that was never promoted, since we
may add tags but never actually take a continuation (a `gather` that
never does `take` would be such an example), so we take care not to
leak those. That duplication can be cleaned up a little later.