Eliminate continuationclone op.
Unused by anything in Rakudo, and in NQP only used in a test case. It
was added because nqp-j had the op, and MoarVM continuations were done
to the same API. However, we never used it besides that test case, and
it didn't really work properly anyway because we didn't need it, so it
is a bit hard to imagine any unknown users seriously relying on it.
The removal of this will help to simplify ->work lifetime management.