Enforce one-shot invocation of continuations.
One-shot is all that Rakudo needs for gather/take, and will suffice
for non-blocking `await` also. We never really properly handled the
multi-shot case anyway, out of not needing it. And our incomplete
attempts to support it came with some nasty memory management risks.
The only non-trivial test for this in the NQP test suite already was
failing, so there's no loss there. This protection will also be useful
for catching bugs when we introduce non-blocking `await` in Rakudo, as
any double invocation there will surely be a screw-up.