Avoid duplicate continuation tag cleanup check.
This was harmless, but a waste of a few cycles on frame destruction.
Given that only frames that were on the call stack will ever have had
opportunity to run a `continuationreset`, there unconditional cleanup
of them in `remove_one_frame` should suffice. (Deserialized contexts
of closures will never have any continuation_tags and no means to get
them.)